Bug 32606: Revert Flatpickr style changes made in Bug 31943
This patch removes the CSS controlling the width of a Flatpickr date
input. The side-effects of this change are worse than the problem the
change was intended to fix.

To test, apply the patch and rebuild the staff interface CSS.

Check pages which are referenced in bugs caused by Bug 31943:

- The holds page (Bug 32259)
- Checking out when specifying a date due in the past (Bug 32210)
- Circulation -> Renew (Bug 32157)

Flatpickr inputs should look correct.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>