Bug 32628: Add 'page-section' to various serials pages
This patch adds a "page-section" container div around the main section
of serials pages which lack it.

This patch contains indentation changes, so please ignore whitespace
when looking at the diff.

To test you must have at least one subscription in your system. Apply
the patch and view the following pages to confirm that the main content
is contained in a white box:

- Subscription batch edit:
  - Go to Serials -> Search subscriptions -> Results
  - Check the box next to one or more search results and click "Edit
    selected serials."
- Subscription frequencies:
  - Serials -> Manage frequencies
- View subscription alerts:
  - Log in to the OPAC and search for a subscription.
    - On the subscription detail page click the "Subscribe to to email
      notification on new issues."
  - In the staff interface, locate that subscription and view its
    details.
    - Click the "subscribers" link under "Patron notification."

Signed-off-by: Andrew Fuerste-Henry <andrewfh@dubcolib.org>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>