Bug 32632: Add 'page-section' to some tools pages
This patch adds a "page-section" container div around the main section
of some tools pages which lack it.

This patch contains indentation changes, so please ignore whitespace
when looking at the diff.

Apply the patch and view the following pages to confirm that the main
content is contained in a white box:

 - From the cataloging home page: Item modifications by age. If
   necessary, add at least one rule to see the table listing all rules.
 - Go to Tools -> Batch extend due dates, and submit values which will
   return one or more results, keeping the "Preview results" checkbox
   checked. Confirm that the page previewing the results looks
   correct.
 - Go to Tools -> Quote editor. If necessary, add at least one quote in
   order to see the table of quotes.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>