Bug 32689: Add missing space between 773$i label text and content
When 777 Ind. 2 = 8, $i will be used as the label. In this case,
we were missing a space between the $i and the other subfields.
This patch makes sure there is a space.

Note: the colon is expected to be part of the data here.

To test:
* Update default framework to display 773 a t w i by default in the editor,
  OPAC and staff interface.
* Search for a record in the staff interface
* Click New > Add child record from the detail page
* Enter 245 and any mandatory fields
* Add $i to 773 entry and make the indicators: 0 8
* On the detail page and the result lists,
  $i will be used a label, but there'll be no space between
  label and content.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit d29057b38262e093d22b47faafeff71da858d408)
Signed-off-by: Jacob O'Mara <jacob.omara@ptfs-europe.com>