Bug 32633: (QA follow-up) Move page-section nesting
This follow-up makes a small change to the page-section nesting on the
Cataloging and Authorities search results pages.  We move the 'h2'
outside of the page section but include the paging and number of results
inside with the results tables themselves.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 734a1fe63ba52b9ce214bb8129493b8184a6f59d)
Signed-off-by: Jacob O'Mara <jacob.omara@ptfs-europe.com>