Bug 32787: Patron restrictions admin page has patron categories search bar
This patch updates the patron restriction administration page to correct
the markup around the header and breadcrumbs: Using the header WRAPPER,
using the sub-header WRAPPER, and switching to the system preferences
search bar.

The patch also corrects the page title to prevent a leading "rsaquo;"
in some cases.

Note: This patch contains indentation changes, so ignore whitespace when
viewing the diff.

To test, apply the patch and go to Administration -> Patron
restrictions. The system preferences search should be preselected in the
header search bar, and everything should look correct.

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>