Bug 31057: Add hint to the 'To date' input on the calendar form
Just adding a hint using the 'hint' class didn't work as
expected here. The hint was immediately hidden. That is
because of the collapsable hint text for the option at
the bottom of the form using the [?] link. So I chose to
use the class of the [$] to make the script a little more
selective.

To test:
* Go to the calendar tool
* Test that the [$] hide/show different help texts
* Apply patch
* Verify a hint now shows below the "To date"
* Verify that the help texts still function as before

Signed-off-by: Matt Blenkinsop <matt.blenkinsop@ptfs-europe.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>