Bug 32941: Sys prefs side menu styling applying where not intended
In bug 32881 I introduced changes to make system preferences sub menu items easier to read. At the time, I couldn't find other instances of ul ul in the templates code, but turns out the ERM module has a <ul> inside a <ul>, and possibly other places.
The original selector was too loose and was applying where not intended, this patch rolls back 32881 changes and puts the changes in a tighter selector.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>